Add support for comma expression operator
(a-la C)

Per chapter 6, challenge 1. This should act basically like C's comma
expression operator, which allows a comma-separated series of
expressions to be given where a single expression is expected (except
inside a function's call list). At runtime, the interpreter will
evaluate the left operand, discard the result, and then evaluate &
return the right operand. It has the lowest precedence of any
operator, and is left-associative.